+++ vllm/entrypoints/openai/serving_engine.py
@@ -1120,9 +1120,7 @@ class OpenAIServing:
         _chat_template_kwargs.update(chat_template_kwargs or {})
 
         request_prompt: str | list[int]
-        print(
-            f"--------{isinstance(tokenizer, DeepseekV32Tokenizer)}--{type(tokenizer)}--"
-        )
+
         if tokenizer is None:
             request_prompt = "placeholder"
         elif isinstance(tokenizer, MistralTokenizer):
@@ -1131,21 +1129,20 @@ class OpenAIServing:
                 messages=messages,
                 **_chat_template_kwargs,
             )
+        elif isinstance(tokenizer, DeepseekV32Tokenizer):
+            request_prompt = tokenizer.apply_chat_template(
+                conversation=conversation,
+                messages=messages,
+                model_config=model_config,
+                **_chat_template_kwargs,
+            )
         else:
-            try:
-                request_prompt = apply_hf_chat_template(
-                    tokenizer=tokenizer,
-                    conversation=conversation,
-                    model_config=model_config,
-                    **_chat_template_kwargs,
-                )
-            except ValueError:
-                request_prompt = tokenizer.apply_chat_template(
-                    conversation=conversation,
-                    messages=messages,
-                    model_config=model_config,
-                    **_chat_template_kwargs,
-                )
+            request_prompt = apply_hf_chat_template(
+                tokenizer=tokenizer,
+                conversation=conversation,
+                model_config=model_config,
+                **_chat_template_kwargs,
+            )
 
         mm_data = await mm_data_future
